datamanager-api/requirements.txt
grpcio-tools==1.51.1
protobuf==4.21.12
uvicorn[standard]
fastapi